% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisaUSA.R
\docType{data}
\name{sch_quest}
\alias{sch_quest}
\title{School-level items for responses in the USA}
\format{Data frame with columns
#' \describe{
  \item{SCHID}{international school ID}
  ...
}}
\source{
http://www.oecd.org/pisa/data/
}
\usage{
sch_quest
}
\description{
School-level items for responses in the USA
}
\keyword{datasets}
