% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisaUSA.R
\docType{data}
\name{stu_quest}
\alias{stu_quest}
\title{All student questionnaire data from the 2015 PISA for students in the United States}
\format{Data frame with columns
#' \describe{
  \item{CNTSTUID}{international student ID}
  \item{SCHID}{international school ID}
  ...
}}
\source{
http://www.oecd.org/pisa/data/
}
\usage{
stu_quest
}
\description{
All student questionnaire data from the 2015 PISA for students in the United States
}
\keyword{datasets}
