% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisaUSA.R
\docType{data}
\name{tch_quest}
\alias{tch_quest}
\title{Items from the PISA 2015 teacher questionnaire for responses in the USA}
\format{Data frame with columns
#' \describe{
  \item{SCHID}{international school ID}
  \item{CNTTCHID}{international teacher ID}
  ...
}}
\source{
http://www.oecd.org/pisa/data/
}
\usage{
tch_quest
}
\description{
Items from the PISA 2015 teacher questionnaire for responses in the USA
}
\keyword{datasets}
